#2 refactor(game-room): move into core
