+++ backend/src/app.module.ts
@@ -3,7 +3,7 @@ import { ServeStaticModule } from '@nestjs/serve-static';
 import { join } from 'path';
 import { AppController } from './app.controller';
 import { MonitoringModule } from './support/monitoring/monitoring.module';
-import { GameRoomModule } from './support/game-room/game-room.module';
+import { GameRoomModule } from './core/game-room/game-room.module';
 
 @Module({
   imports: [
