+++ backend/src/core/game-room/game-room.module.ts
